+++ packages/sdk/src/services/lp/commonSimulateLp.ts
